requirements.txt
torch
allennlp==1.0.0
stanza
tqdm
transformers
PyYAML
click
h5py
python-Levenshtein
matplotlib
numpy
seaborn
